Move helper to GUI unit test namespace
