+++ tests/unit/cmk/gui/plugins/metrics/test_unit.py
@@ -8,7 +8,7 @@ from typing import ContextManager
 
 import pytest
 
-from tests.testlib.users import create_and_destroy_user
+from tests.unit.cmk.gui.users import create_and_destroy_user
 
 from cmk.utils.user import UserId
