+++ tests/unit/cmk/gui/test_login_unit.py
@@ -14,9 +14,8 @@ import flask
 import pytest
 from werkzeug.test import create_environ
 
-from tests.testlib.users import create_and_destroy_user
-
 from tests.unit.cmk.gui.conftest import WebTestAppForCMK
+from tests.unit.cmk.gui.users import create_and_destroy_user
 
 from cmk.utils.livestatus_helpers.testing import MockLiveStatusConnection
 from cmk.utils.user import UserId
